+++ bindings/vue/examples/components/NavigatorVuex.vue
@@ -1,6 +1,6 @@
 <template>
   <v-ons-page>
-    <v-ons-navigator :page-stack="pageStack" :pop-page="pop"></v-ons-navigator>
+    <v-ons-navigator swipeable :page-stack="pageStack" :pop-page="pop"></v-ons-navigator>
   </v-ons-page>
 </template>
 
@@ -77,13 +77,9 @@
     methods: {
       push() {
         this.$store.commit('pushPage', [page3, page3, page3]);
-        this.$nextTick(() => this.navigator.isReady().then(() => {
-          console.log('is ready');
-        }));
       }
     },
     components: { myToolbar },
-    inject: ['navigator'],
     mounted() {
     }
   };