+++ bindings/vue/src/mixins/common.js
@@ -44,17 +44,6 @@ const hidable = {
   }
 };
 
-// Clickable
-const simpleClick = {
-  mounted() {
-    this._boundOnClick = e => this.$emit('click', e);
-    this.$el.addEventListener('click', this._boundOnClick);
-  },
-  beforeDestroy() {
-    this.$el.removeEventListener('click', this._boundOnClick);
-  }
-};
-
 // Components with 'options' property
 const hasOptions = {
   props: {
@@ -139,4 +128,4 @@ const portal = {
   }
 };
 
-export { hidable, simpleClick, hasOptions, modifier, selfProvider, dialogCancel, portal };
+export { hidable, hasOptions, modifier, selfProvider, dialogCancel, portal };