+++ bindings/vue/src/mixins/derive.js
@@ -1,4 +1,4 @@
-import { eventToHandler, handlerToProp } from '../internal/util';
+import { capitalize, camelize, eventToHandler, handlerToProp } from '../internal/util';
 
 /* Private */
 const _setupDBB = component => {
@@ -67,6 +67,18 @@ const deriveHandler = handlerName => {
 };
 
 const deriveEvents = {
+  computed: {
+    unrecognizedListeners() {
+      const element = capitalize(camelize(this.$options._componentTag.slice(6))) + 'Element';
+      return Object.keys(this.$listeners || {})
+        .filter(k => (this.$ons[element].events || []).indexOf(k) === -1)
+        .reduce((r, k) => {
+          r[k] = this.$listeners[k];
+          return r;
+        }, {});
+    }
+  },
+
   mounted() {
     this._handlers = {};