+++ core/src/elements/ons-alert-dialog/index.js
@@ -30,10 +30,11 @@ const scheme = {
   '.alert-dialog-title': 'alert-dialog-title--*',
   '.alert-dialog-content': 'alert-dialog-content--*',
   '.alert-dialog-footer': 'alert-dialog-footer--*',
-  '.alert-dialog-button': 'alert-dialog-button--*',
   '.alert-dialog-footer--rowfooter': 'alert-dialog-footer--rowfooter--*',
   '.alert-dialog-button--rowfooter': 'alert-dialog-button--rowfooter--*',
   '.alert-dialog-button--primal': 'alert-dialog-button--primal--*',
+  '.alert-dialog-button': 'alert-dialog-button--*',
+  'ons-alert-dialog-button': 'alert-dialog-button--*',
   '.alert-dialog-mask': 'alert-dialog-mask--*',
   '.text-input': 'text-input--*'
 };